+++ src/store/settings.ts
@@ -82,6 +82,9 @@ export const checkUpgradeCore = useStorage('config/check-upgrade-core', true)
 export const autoUpgradeCore = useStorage('config/auto-upgrade-core', false)
 export const swipeInTabs = useStorage('config/swipe-in-tabs', false)
 export const disablePullToRefresh = useStorage('config/disable-pull-to-refresh', true)
+export const displayAllFeatures = useStorage('config/display-all-features', false)
+export const blurIntensity = useStorage('config/blur-intensity', 10)
+export const scrollAnimationEffect = useStorage('config/scroll-animation-effect', true)
 
 // overview
 export const splitOverviewPage = useStorage('config/split-overview-page', false)
@@ -92,6 +95,11 @@ export const showStatisticsWhenSidebarCollapsed = useStorage(
   'config/show-statistics-when-sidebar-collapsed',
   true,
 )
+export const numberOfChartsInSidebar = useStorage<1 | 2 | 3>(
+  'config/number-of-charts-in-sidebar',
+  2,
+)
+
 // proxies
 export const collapseGroupMap = useStorage<Record<string, boolean>>('config/collapse-group-map', {})
 export const twoColumnProxyGroup = useStorage('config/two-columns', true)